+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * Plugin files for icon and banner 
 * Update FR translation
 * Check $contact_page_id is defined
+* Save Aquos ID in order data for regular sale process (not appointments)
 
 ### 1.1.7: July 28th, 2020
 * New setting field to specify excluded WooCommerce product categories